Move sendNotification and sendPledgeNotification to conversationAndEmailHandler, define input for dappmailer functions
