+++ src/utils/dappMailer.js
@@ -43,19 +43,23 @@ const sendEmail = (app, data) => {
 };
 
 module.exports = {
-  donation: (app, data) => {
-    data.amount = Number(data.amount) / 10 ** 18;
+  thanksFromDonationGiver: (
+    app,
+    { recipient, user, amount, token, donationType, donatedToTitle },
+  ) => {
+    const normalizedAmount = Number(amount) / 10 ** 18;
 
-    Object.assign(data, {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Thank you for your donation!',
-      secretIntro: `Thank you for your donation of ${data.amount} ${data.token.symbol} to the ${data.donationType} "${data.donatedToTitle}"!`,
+      secretIntro: `Thank you for your donation of ${normalizedAmount} ${token.symbol} to the ${donationType} "${donatedToTitle}"!`,
       title: 'You are so awesome!',
       image: 'Giveth-donation-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          Thank you very much for your donation of ${data.amount} ${data.token.symbol} to the ${data.donationType} <em>${data.donatedToTitle}</em>.
+          Thank you very much for your donation of ${normalizedAmount} ${token.symbol} to the ${donationType} <em>${donatedToTitle}</em>.
           With your donation we can really make this happen, and you play a vital part in making the world a better place!
         </p>
       `,
@@ -63,95 +67,121 @@ module.exports = {
       ctaRelativeUrl: '/donations',
       unsubscribeType: 'donation-receipt',
       unsubscribeReason: 'You receive this email from Giveth because you have made a donation',
-    });
+    };
 
     sendEmail(app, data);
   },
 
-  donationReceived: (app, data) => {
-    data.amount = Number(data.amount) / 10 ** 18;
-    Object.assign(data, {
+  donationReceived: (app, { recipient, user, donationType, donatedToTitle, amount, token }) => {
+    const normalizedAmount = Number(amount) / 10 ** 18;
+    const data = {
+      recipient,
       template: 'notification',
       subject: "Giveth - You've received a donation!",
-      secretIntro: `You have received a donation of ${data.amount} ${data.token.symbol} for the ${data.donationType} "${data.donatedToTitle}"!`,
+      secretIntro: `You have received a donation of ${normalizedAmount} ${token.symbol} for the ${donationType} "${donatedToTitle}"!`,
       title: 'You are so awesome!',
       image: 'Giveth-donation-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
           You have received a donation of
-          <span>${data.amount} ${data.token.symbol}</span>
-          for your ${data.donationType} <em>${data.donatedToTitle}</em>.
+          <span>${amount} ${token.symbol}</span>
+          for your ${donationType} <em>${donatedToTitle}</em>.
         </p>
       `,
-      cta: `Manage your ${data.donationType}`,
-      ctaRelativeUrl: `/my-${data.donationType}s`,
+      cta: `Manage your ${donationType}`,
+      ctaRelativeUrl: `/my-${donationType}s`,
       unsubscribeType: 'donation-received',
-      unsubscribeReason: `You receive this email because you run a ${data.donationType}`,
-    });
+      unsubscribeReason: `You receive this email because you run a ${donationType}`,
+    };
 
     sendEmail(app, data);
   },
 
-  delegationRequired: (app, data) => {
-    data.amount = Number(data.amount) / 10 ** 18;
-
-    Object.assign(data, {
+  delegationRequired: (
+    app,
+    {
+      recipient,
+      user,
+      donationType, // dac / campaign
+      donatedToTitle,
+      amount,
+      token,
+    },
+  ) => {
+    const normalizedAmount = Number(amount) / 10 ** 18;
+    const data = {
+      recipient,
+      user,
       template: 'notification',
       subject: 'Giveth - Delegation required for new donation!',
-      secretIntro: `Take action! Please delegate a new donation of ${data.amount} ${data.token.symbol} for the ${data.donationType} "${data.donatedToTitle}"!`,
+      secretIntro: `Take action! Please delegate a new donation of ${normalizedAmount} ${token.symbol} for the ${donationType} "${donatedToTitle}"!`,
       title: "Take action! You've received a donation, delegate now!",
       image: 'Giveth-donation-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
           You have received a donation of
-          <span style="display: block; color: rgb(53, 184, 209); line-height: 72px; font-size: 48px;">${
-            data.amount
-          } ${data.token.symbol}</span>
-          for your ${data.donationType} <em>${data.donatedToTitle}</em>.
+          <span style="display: block; color: rgb(53, 184, 209); line-height: 72px; font-size: 48px;">${amount} ${
+        token.symbol
+      }</span>
+          for your ${donationType} <em>${donatedToTitle}</em>.
         </p>
         <p>
           You can now delegate this money to a ${
-            data.donationType === AdminTypes.DAC ? 'Campaign or a Milestone' : 'Milestone'
+            donationType === AdminTypes.DAC ? 'Campaign or a Milestone' : 'Milestone'
           }.
         </p>
       `,
       cta: `Delegate Donation`,
       ctaRelativeUrl: `/delegations`,
       unsubscribeType: 'request-delegation',
-      unsubscribeReason: `You receive this email because you run a ${data.donationType}`,
-    });
+      unsubscribeReason: `You receive this email because you run a ${donationType}`,
+    };
 
     sendEmail(app, data);
   },
 
-  donationDelegated: (app, data) => {
-    data.amount = Number(data.amount) / 10 ** 18;
+  donationDelegated: (
+    app,
+    {
+      recipient,
+      user,
+      delegationType,
+      delegatedToTitle,
+      delegateType,
+      delegateTitle,
+      commitTime,
+      amount,
+      token,
+    },
+  ) => {
+    const normalizedAmount = Number(amount) / 10 ** 18;
 
-    const capitalizeDelegateType = delegateType => {
-      if (delegateType.toLowerCase() === 'dac') return 'DAC';
-      return delegateType.charAt(0).toUpperCase() + delegateType.slice(1);
+    const capitalizeDelegateType = inputDelegateType => {
+      if (inputDelegateType.toLowerCase() === 'dac') return 'DAC';
+      return inputDelegateType.charAt(0).toUpperCase() + inputDelegateType.slice(1);
     };
 
-    Object.assign(data, {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Your donation has been delegated!',
-      secretIntro: `Take action! Please approve or reject the delegation of ${data.amount} ${data.token.symbol} to the ${data.delegationType} "${data.delegatedToTitle}"!`,
+      secretIntro: `Take action! Please approve or reject the delegation of ${normalizedAmount} ${token.symbol} to the ${delegationType} "${delegatedToTitle}"!`,
       title: 'Take action! Your donation has been delegated!',
       image: 'Giveth-donation-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          The ${capitalizeDelegateType(data.delegateType)} <em>${
-        data.delegateTitle
-      }</em> has proposed a delegation of
+          The ${capitalizeDelegateType(
+            delegateType,
+          )} <em>${delegateTitle}</em> has proposed a delegation of
           <span style="display: block; color: rgb(53, 184, 209); line-height: 72px; font-size: 48px;">
-          ${data.amount} ${data.token.symbol}</span> from your donation to
-          ${capitalizeDelegateType(data.delegateType)} <em>${data.delegateTitle}</em>.
+          ${normalizedAmount} ${token.symbol}</span> from your donation to
+          ${capitalizeDelegateType(delegateType)} <em>${delegateTitle}</em>.
         </p>
         <p>
-          You have until ${data.commitTime.toUTCString()} to approve or reject this delegation. If you fail to
+          You have until ${commitTime.toUTCString()} to approve or reject this delegation. If you fail to
           act before this date, this delegation will be auto-approved.
         </p>
       `,
@@ -159,99 +189,115 @@ module.exports = {
       ctaRelativeUrl: `/donations`,
       unsubscribeType: 'donation-delegated',
       unsubscribeReason: `You receive this email because your donation was delegated`,
-    });
+    };
 
     sendEmail(app, data);
   },
 
-  milestoneProposed: (app, data) => {
-    data.amount = Number(data.amount) / 10 ** 18;
+  milestoneProposed: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignTitle, campaignId, amount, token },
+  ) => {
+    const normalizedAmount = Number(amount) / 10 ** 18;
 
-    Object.assign(data, {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - A Milestone has been proposed!',
       secretIntro: `Take action! A Milestone has been proposed for your Campaign! Please accept or reject.`,
       title: 'Take action: Milestone proposed!',
       image: 'Giveth-suggest-milestone-banner.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          The Milestone <em>${data.milestoneTitle}</em> for <em>${data.amount} ${data.token.symbol}</em> has been proposed to <em>${data.campaignTitle}</em> Campaign .
+          The Milestone <em>${milestoneTitle}</em> for <em>${normalizedAmount} ${token.symbol}</em> has been proposed to <em>${campaignTitle}</em> Campaign .
           If you think this is a great idea, then <strong>please approve this Milestone within 3 days</strong> to add it to your Campaign.
           If not, then please reject it with comment.
         </p>
       `,
       cta: `See the Milestone`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'milestone-proposed',
       unsubscribeReason: `You receive this email because you run a Campaign`,
-      message: data.message,
-    });
+      // message: message,
+    };
 
     sendEmail(app, data);
   },
 
-  proposedMilestoneAccepted: (app, data) => {
-    Object.assign(data, {
+  proposedMilestoneAccepted: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignTitle, campaignId, message },
+  ) => {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Your proposed Milestone is accepted!',
-      secretIntro: `Your Milestone ${data.milestoneTitle} has been accepted by the Campaign Owner. You can now receive donations.`,
+      secretIntro: `Your Milestone ${milestoneTitle} has been accepted by the Campaign Owner. You can now receive donations.`,
       title: 'Take action: Milestone proposed!',
       image: 'Giveth-milestone-review-approved-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          Your proposed Milestone <em>${data.milestoneTitle}</em> to the Campaign <em>${data.campaignTitle}</em> has been accepted by the Campaign Owner!
+          Your proposed Milestone <em>${milestoneTitle}</em> to the Campaign <em>${campaignTitle}</em> has been accepted by the Campaign Owner!
           <br/><br/>
           You can now receive donations, start executing the Milestone, and once finished, mark it as complete.
         </p>
       `,
       cta: `Manage Milestone`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'proposed-milestone-accepted',
       unsubscribeReason: `You receive this email because you run a Milestone`,
-      message: data.message,
-    });
+      message,
+    };
 
     sendEmail(app, data);
   },
 
-  proposedMilestoneRejected: (app, data) => {
-    Object.assign(data, {
+  proposedMilestoneRejected: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignTitle, campaignId, message },
+  ) => {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Your proposed Milestone is rejected :-(',
-      secretIntro: `Your Milestone ${data.milestoneTitle} has been rejected by the Campaign Owner :-(`,
+      secretIntro: `Your Milestone ${milestoneTitle} has been rejected by the Campaign Owner :-(`,
       title: 'Milestone rejected :-(',
       image: 'Giveth-milestone-review-approved-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          Unfortunately your proposed Milestone <em>${data.milestoneTitle}</em> to the Campaign <em>${data.campaignTitle}</em> has been rejected by the Campaign Owner.
+          Unfortunately your proposed Milestone <em>${milestoneTitle}</em> to the Campaign <em>${campaignTitle}</em> has been rejected by the Campaign Owner.
           <br/><br/>
           Please contact the Campaign Owner to learn why your Milestone was rejected.
         </p>
       `,
       cta: `Manage Milestone`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'proposed-milestone-rejected',
       unsubscribeReason: `You receive this email because you proposed a Milestone`,
-      message: data.message,
-    });
+      message,
+    };
 
     sendEmail(app, data);
   },
 
-  milestoneRequestReview: (app, data) => {
-    Object.assign(data, {
+  milestoneRequestReview: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignTitle, campaignId, message },
+  ) => {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Time to review!',
-      secretIntro: `Take action: you are requested to review the Milestone ${data.milestoneTitle} within 3 days.`,
+      secretIntro: `Take action: you are requested to review the Milestone ${milestoneTitle} within 3 days.`,
       title: 'Milestone review requested',
       image: 'Giveth-review-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          The Milestone <em>${data.milestoneTitle}</em> to the Campaign <em>${data.campaignTitle}</em> has been marked as completed by the Milestone Owner.
+          The Milestone <em>${milestoneTitle}</em> to the Campaign <em>${campaignTitle}</em> has been marked as completed by the Milestone Owner.
           <br/><br/>
         </p>
           Now is your moment to shine!
@@ -261,151 +307,169 @@ module.exports = {
         </p>
       `,
       cta: `Review Milestone`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'milestone-request-review',
       unsubscribeReason: `You receive this email because you run a Milestone`,
-      message: data.message,
-    });
+      message,
+    };
 
     sendEmail(app, data);
   },
 
-  milestoneMarkedCompleted: (app, data) => {
-    Object.assign(data, {
+  milestoneMarkedCompleted: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignTitle, campaignId, message, token },
+  ) => {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Your Milestone is finished!',
-      secretIntro: `Your Milestone ${data.milestoneTitle} has been marked complete by the reviewer. The recipient can now collect the payment.`,
-      title: `Milestone completed! Time to collect ${data.token.symbol}.`,
+      secretIntro: `Your Milestone ${milestoneTitle} has been marked complete by the reviewer. The recipient can now collect the payment.`,
+      title: `Milestone completed! Time to collect ${token.symbol}.`,
       image: 'Giveth-milestone-review-approved-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          The Milestone <em>${data.milestoneTitle}</em> in the Campaign <em>${data.campaignTitle}</em> has been marked complete by the reviewer!.
+          The Milestone <em>${milestoneTitle}</em> in the Campaign <em>${campaignTitle}</em> has been marked complete by the reviewer!.
           <br/><br/>
         </p>
           The recipient can now transfer the funds out of this Milestone!
         </p>
       `,
       cta: `Manage Milestone`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'milestone-review-approved',
       unsubscribeReason: `You receive this email because you run a Milestone`,
-      message: data.message,
-    });
+      message,
+    };
 
     sendEmail(app, data);
   },
 
-  milestoneReviewRejected: (app, data) => {
-    Object.assign(data, {
+  milestoneReviewRejected: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignTitle, campaignId, message },
+  ) => {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Milestone rejected by reviewer :-(',
       type: 'milestone-review-rejected',
-      secretIntro: `The completion of your Milestone ${data.milestoneTitle} has been rejected by the reviewer.`,
+      secretIntro: `The completion of your Milestone ${milestoneTitle} has been rejected by the reviewer.`,
       title: 'Milestone completion rejected.',
       image: 'Giveth-milestone-review-rejected-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          The Milestone completion <em>${data.milestoneTitle}</em> in the Campaign <em>${data.campaignTitle}</em> has been rejected by the reviewer.
+          The Milestone completion <em>${milestoneTitle}</em> in the Campaign <em>${campaignTitle}</em> has been rejected by the reviewer.
         </p>
       `,
       cta: `Manage Milestone`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'milestone-review-rejected',
       unsubscribeReason: `You receive this email because you run a Milestone`,
-      message: data.message,
-    });
+      message,
+    };
 
     sendEmail(app, data);
   },
 
-  milestoneCanceled: (app, data) => {
-    Object.assign(data, {
+  milestoneCanceled: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignTitle, campaignId, message },
+  ) => {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Milestone canceled :-(',
       type: 'milestone-canceled',
-      secretIntro: `Your Milestone ${data.milestoneTitle} has been canceled.`,
+      secretIntro: `Your Milestone ${milestoneTitle} has been canceled.`,
       title: 'Milestone Canceled',
       image: 'Giveth-milestone-canceled-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          The Milestone <em>${data.milestoneTitle}</em> in the Campaign <em>${data.campaignTitle}</em> has been canceled.
+          The Milestone <em>${milestoneTitle}</em> in the Campaign <em>${campaignTitle}</em> has been canceled.
         </p>
       `,
       cta: `Manage Milestones`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'milestone-canceled',
       unsubscribeReason: `You receive this email because you run a Milestone`,
-      message: data.message,
-    });
+      message,
+    };
 
     sendEmail(app, data);
   },
 
-  milestoneCreated: (app, data) => {
-    data.amount = Number(data.amount) / 10 ** 18;
+  milestoneCreated: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignId, amount, token },
+  ) => {
+    const normalizedAmount = Number(amount) / 10 ** 18;
 
-    Object.assign(data, {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Milestone created with you as a recipient',
       type: 'milestone-created',
-      secretIntro: `A Milestone ${data.milestoneTitle} has been created with you as the recipient.`,
+      secretIntro: `A Milestone ${milestoneTitle} has been created with you as the recipient.`,
       title: 'Milestone created.',
       image: 'Giveth-milestone-review-approved-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
         <p>
-          A Milestone <em>${data.milestoneTitle}</em> for ${data.amount} ${data.token.symbol} has been created with you as the recipient.
+          A Milestone <em>${milestoneTitle}</em> for ${normalizedAmount} ${token.symbol} has been created with you as the recipient.
         </p>
       `,
       cta: `See your Milestones`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'milestone-created',
       unsubscribeReason: `You receive this email because you are the recipient of a Milestone`,
-      message: data.message,
-    });
+    };
 
     sendEmail(app, data);
   },
 
-  donationCancelled: (app, data) => {
-    Object.assign(data, {
-      subject: 'Giveth - Oh no, you lost a giver!',
-      type: 'donation-cancelled',
-    });
+  // This wasn't used so I commented that
+  // donationCancelled: (app, data) => {
+  //   Object.assign(data, {
+  //     subject: 'Giveth - Oh no, you lost a giver!',
+  //     type: 'donation-cancelled',
+  //   });
+  //
+  //   // not implemented yet
+  //   // sendEmail(app, data);
+  //   sendEmail(app, data);
+  // },
 
-    // not implemented yet
-    // sendEmail(app, data);
-    sendEmail(app, data);
-  },
-
-  donationsCollected: (app, data) => {
-    Object.assign(data, {
+  donationsCollected: (
+    app,
+    { recipient, user, milestoneTitle, milestoneId, campaignId, conversation, address },
+  ) => {
+    const data = {
+      recipient,
       template: 'notification',
       subject: 'Giveth - Donations collected',
       type: 'milestone-donations-collected',
-      secretIntro: `Your Milestone ${data.milestoneTitle} has been paid.`,
+      secretIntro: `Your Milestone ${milestoneTitle} has been paid.`,
       title: 'Milestone Donations Collected',
       image: 'Giveth-milestone-review-approved-banner-email.png',
       text: `
-        <p><span style="line-height: 33px; font-size: 22px;">Hi ${data.user}</span></p>
-        <p>The following payments have been initiated for your Milestone <em>${
-          data.milestoneTitle
-        }</em>:</p>
+        <p><span style="line-height: 33px; font-size: 22px;">Hi ${user}</span></p>
+        <p>The following payments have been initiated for your Milestone <em>${milestoneTitle}</em>:</p>
         <p></p>
-        ${data.conversation.payments.map(p => `<p>${p.amount / 10 ** 18} ${p.symbol}</p>`)}
+        ${conversation.payments.map(p => `<p>${p.amount / 10 ** 18} ${p.symbol}</p>`)}
         <p></p>
-        <p>You can expect to see these payment(s) to arrive in your wallet <em>${
-          data.address
-        }</em> within 48 - 72 hrs.</p>
+        <p>You can expect to see these payment(s) to arrive in your wallet <em>
+           ${address}
+        </em> within 48 - 72 hrs.</p>
       `,
       cta: `See your Milestones`,
-      ctaRelativeUrl: `/campaigns/${data.campaignId}/milestones/${data.milestoneId}`,
+      ctaRelativeUrl: `/campaigns/${campaignId}/milestones/${milestoneId}`,
       unsubscribeType: 'donations-collected',
       unsubscribeReason: `You receive this email because you are the recipient of a Milestone`,
-    });
+    };
     sendEmail(app, data);
   },
 };
